add interface to fix proxy jdk issue
